An application that uses Shamir's Secret Sharing Scheme on image pixels to obfuscate biometric image data
This was my final project for my Data Privacy and Security class.  One of the cryptographic techniques that we studied was Shamir's Secret Sharing Scheme.  The (t,n) sharing scheme splits a secret up into n unique  pieces that contain information to "reconstruct" the secret.  The secret can only be reconstructed if information from at least t shares are combined.  In this application, pixels from an image are encrypted using the scheme, resulting in n new images.  By combining t of these n images, the original image can be restored.